requirements.txt
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
transformers>=4.39.0
accelerate>=0.28.0
huggingface-hub<0.27
pandas
numpy<2.0
librosa
soundfile
pyannote.audio==3.3.1
python-dotenv
ipykernel
jupyter